Adding csp with maptiler/osm origins, adding offline maps configuratoin
