+++ infrastructure/aws-cdk/configs/sample.json
@@ -21,7 +21,13 @@
     "notebookListType": "tabs",
     "notebookName": "notebook",
     "appName": "FAIMS Mobile",
-    "appId": "FAIMS"
+    "appId": "FAIMS",
+    "offlineMaps": {
+      "mapSource": "maptiler",
+      "mapSourceKey": "your-maptiler-api-key",
+      "offlineMaps": true,
+      "mapStyle": "basic"
+    }
   },
   "couch": {
     "ebsRecoverySnapshotId": "1234",
